fix bug getting user like status, comment date formatter console.log, post like revision from using two endpoint to one enpoint on create on delete like, fix getting avatar image on route on server.js

+++ services/PostLikeService.js
@@ -27,36 +27,56 @@ class PostLikeService {
   }
 
   // Create a new post like
-  static async create(postId, userId) {
+  static async createDelete(postId, userId) {
     try {
+      // Check if post like already exists
+      const existingLike = await PostLikeRepository.findByUserIdAndPostId(postId, userId);
+
+      if (existingLike) {
+        // If it exists, delete the like
+        await PostLikeRepository.delete(postId, userId);
+        return { message: "Post like deleted" };
+      }
+
       // Create a new post like
       const postLike = await PostLikeRepository.create(postId, userId);
-      if (!postLike) {
-        throw new Error(`${currentService} Error: Post like not found`);
-      }
 
-      // Return post like
-      return postLike;
+      // Check if post like was created successfully
+      if (!postLike) {
+        throw new Error(`${currentService} Error: Post like cannot be created`);
+      } else {
+        // If it was created successfully, return the post like
+        const postLikeData = {
+          id: postLike.id,
+          user_id: postLike.user_id,
+          post_id: postLike.post_id,
+          created_at: postLike.created_at,
+          updated_at: postLike.updated_at,
+        };
+
+        // Return success message with post like data
+        return { message: "Post like created", data: postLikeData };
+      }
     } catch (error) {
       throw new Error(`${currentService} Error: ${error.message}`);
     }
   }
 
-  // Delete a post like
-  static async delete(postId, userId) {
-    try {
-      // Delete a post like by post id
-      const postLike = await PostLikeRepository.delete(postId, userId);
-      if (!postLike) {
-        throw new Error(`${currentService} Error: Post ID not found`);
-      }
+  // // Delete a post like
+  // static async delete(postId, userId) {
+  //   try {
+  //     // Delete a post like by post id
+  //     const postLike = await PostLikeRepository.delete(postId, userId);
+  //     if (!postLike) {
+  //       throw new Error(`${currentService} Error: Post ID not found`);
+  //     }
 
-      // Return deleted post like
-      return postLike;
-    } catch (error) {
-      throw new Error(`${currentService} Error: ${error.message}`);
-    }
-  }
+  //     // Return deleted post like
+  //     return postLike;
+  //   } catch (error) {
+  //     throw new Error(`${currentService} Error: ${error.message}`);
+  //   }
+  // }
 }
 
 module.exports = PostLikeService;
